requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-dotenv
pillow